requirements.txt
# Praeto Tender Tracker - Python Dependencies
# Install: pip install -r requirements.txt

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Google Sheets
gspread==5.12.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0

# Data Processing
pandas
numpy

# Dashboard
streamlit
plotly
xlsxwriter

# Email
secure-smtplib==0.1.1

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
